Fixes #413, whereby the NamedTemporaryFile "dest" was out of scope in the error handling portion of rewrite.

+++ src/dotenv/main.py
@@ -125,15 +125,16 @@ def rewrite(
     path: Union[str, os.PathLike],
     encoding: Optional[str],
 ) -> Iterator[Tuple[IO[str], IO[str]]]:
+    dest = None
     try:
         if not os.path.isfile(path):
             with open(path, "w+", encoding=encoding) as source:
                 source.write("")
-        with tempfile.NamedTemporaryFile(mode="w+", delete=False, encoding=encoding) as dest:
-            with open(path, encoding=encoding) as source:
-                yield (source, dest)  # type: ignore
+        dest = tempfile.NamedTemporaryFile(mode="w+", delete=False, encoding=encoding)
+        with open(path, encoding=encoding) as source:
+            yield (source, dest)  # type: ignore
     except BaseException:
-        if os.path.isfile(dest.name):
+        if dest and os.path.isfile(dest.name):
             os.unlink(dest.name)
         raise
     else:
